requirements.txt
# Core RAG dependencies
llama-index==0.10.0
llama-index-embeddings-huggingface==0.2.0
llama-index-vector-stores-chroma==0.1.6
llama-index-llms-groq==0.1.3

# Vector store
chromadb==0.4.22

# LLM & Embeddings
groq==0.4.2
sentence-transformers==2.3.1

# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0
trafilatura==1.6.3

# UI
gradio==4.16.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1

# Development (optional)
pytest==7.4.4
black==24.1.1
flake8==7.0.0